requirements.txt
# QuantMindX Python Dependencies

# Web Scraping
firecrawl-py>=0.0.16
finnhub-python>=2.4.0  # Finnhub news feed provider (Story 6.3)
requests>=2.31.0
beautifulsoup4>=4.12.0

# HTTP Client for PageIndex
httpx>=0.24.0

# SSH for server connectivity testing
paramiko>=3.0.0

# LLM Integration (v2 Agent Stack)
# LangChain/LangGraph dependencies removed - now using Claude CLI directly
# langchain>=0.1.0  # REMOVED - migrated to claude-cli
# langchain-openai>=0.1.0  # REMOVED
# langchain-community>=0.0.10  # REMOVED
# langgraph>=0.2.0  # REMOVED
# langmem>=0.0.30  # REMOVED - migrated to claude-mem, now removed entirely
# langchain-core>=0.1.0  # REMOVED
# langgraph-checkpoint>=0.0.9  # REMOVED
openai>=1.0.0  # Kept for legacy compatibility
anthropic>=0.5.0  # Required for Claude API

# CLI Framework
click>=8.1.0
rich>=13.0.0
pyfiglet>=1.0.0
textual>=0.47.0
psutil>=5.9.0
watchdog>=3.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.10.0
pydantic>=2.0.0
pyyaml>=6.0
duckdb>=0.9.0
pyarrow>=14.0.0

# MCP Server
mcp>=1.0.0
-e ./mcp-metatrader5-server

# NotebookLM MCP for knowledge management
notebooklm-mcp-cli>=0.1.0

# Workflow Orchestration - Prefect for Alpha Forge pipeline
prefect>=2.20.0

# Utilities
python-dotenv>=1.0.0
cryptography>=3.4.0
tqdm>=4.66.0
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.2.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# V8 Expansion Dependencies
pyzmq>=25.1.2          # ZMQ socket support for HFT infrastructure
aiohttp>=3.9.1         # Async HTTP for WebSocket (Crypto module)
ccxt>=4.2.0            # Optional: Unified exchange API

# Memory & Embeddings
sentence-transformers>=2.2.0  # all-MiniLM-L6-v2 for vector embeddings
chromadb>=0.4.0        # Vector database for embedding storage

# Monitoring & Observability
prometheus-client==0.19.0
python-snappy>=0.6.0        # Snappy compression for Grafana Cloud remote_write
prometheus-remote-write>=0.1.0  # Protobuf WriteRequest encoding for remote_write

# OpenTelemetry for Distributed Tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-sdk-extension-aws>=1.4.0  # AWS X-Ray support (optional)
